admm-logistic-regression/requirements.txt
matplotlib
torchnet
